+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsUser/MSFT_TeamsUser.psm1
@@ -184,13 +184,13 @@ function Set-TargetResource
     Write-Verbose -Message "Retrieve team GroupId: $($team.GroupId)"
 
     $CurrentParameters = $PSBoundParameters
-    $CurrentParameters.Remove("TeamName")
+    $CurrentParameters.Remove("TeamName") | Out-Null
     $CurrentParameters.Add("GroupId", $team.GroupId)
-    $CurrentParameters.Remove("GlobalAdminAccount")
-    $CurrentParameters.Remove("ApplicationId")
-    $CurrentParameters.Remove("TenantId")
-    $CurrentParameters.Remove("CertificateThumbprint")
-    $CurrentParameters.Remove("Ensure")
+    $CurrentParameters.Remove("GlobalAdminAccount") | Out-Null
+    $CurrentParameters.Remove("ApplicationId") | Out-Null
+    $CurrentParameters.Remove("TenantId") | Out-Null
+    $CurrentParameters.Remove("CertificateThumbprint") | Out-Null
+    $CurrentParameters.Remove("Ensure") | Out-Null
 
     if ($Ensure -eq "Present")
     {
@@ -201,7 +201,7 @@ function Set-TargetResource
     {
         if ($Role -eq "Member" -and $CurrentParameters.ContainsKey("Role"))
         {
-            $CurrentParameters.Remove("Role")
+            $CurrentParameters.Remove("Role") | Out-Null
             Write-Verbose -Message "Removed role parameter"
         }
         Remove-TeamUser @CurrentParameters
@@ -265,7 +265,7 @@ function Test-TargetResource
 
     if ($null -eq $Role)
     {
-        $CurrentValues.Remove("Role")
+        $CurrentValues.Remove("Role") | Out-Null
     }
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
@@ -400,7 +400,7 @@ function Export-TargetResource
                 $returnValue += Invoke-M365DSCCommand -Arguments $PSBoundParameters -InvokationPath $ScriptRoot -ScriptBlock {
                     $WarningPreference = 'SilentlyContinue'
                     $params = $args[0]
-                    $content = ""
+                    $dscContent = ""
                     $j = 1
                     $ConnectionMode = New-M365DSCConnection -Platform 'MicrosoftTeams' -InboundParameters $PSBoundParameters
 
@@ -452,31 +452,34 @@ function Export-TargetResource
                                     if ($ConnectionMode -eq 'Credential')
                                     {
                                         $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
-                                        $result.Remove("ApplicationId")
-                                        $result.Remove("TenantId")
-                                        $result.Remove("CertificateThumbprint")
+                                        $result.Remove("ApplicationId") | Out-Null
+                                        $result.Remove("TenantId") | Out-Null
+                                        $result.Remove("CertificateThumbprint") | Out-Null
                                     }
                                     else
                                     {
-                                        $result.Remove("GlobalAdminAccount")
+                                        $result.Remove("GlobalAdminAccount") | Out-Null
                                     }
-                                    $content += "        TeamsUser " + (New-Guid).ToString() + "`r`n"
-                                    $content += "        {`r`n"
-                                    $currentDSCBlock = Get-DSCBlock -Params $result -ModulePath $params.ScriptRoot
+                                    $currentDSCBlock = "        TeamsUser " + (New-Guid).ToString() + "`r`n"
+                                    $currentDSCBlock += "        {`r`n"
+                                    $content = Get-DSCBlock -Params $result -ModulePath $params.ScriptRoot
                                     if ($ConnectionMode -eq 'Credential')
                                     {
-                                        $partialContent = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
+                                        $partialContent = Convert-DSCStringParamToVariable -DSCBlock $content -ParameterName "GlobalAdminAccount"
                                     }
                                     else
                                     {
-                                        $partialContent = $currentDSCBlock
+                                        $partialContent = $content
                                     }
                                     $partialContent += "        }`r`n"
                                     if ($partialContent.ToLower().Contains($params.OrganizationName.ToLower()))
                                     {
                                         $partialContent = $partialContent -ireplace [regex]::Escape($params.OrganizationName), "`$OrganizationName"
                                     }
-                                    $content += $partialContent
+                                    $currentDSCBlock += $partialContent
+                                    $dscContent += $currentDSCBlock
+                                    Save-M365DSCPartialExport -Content $currentDSCBlock `
+                                        -FileName $Global:PartialExportFileName
                                     $i++
                                 }
                             }
@@ -488,7 +491,7 @@ function Export-TargetResource
                             $j++
                         }
                     }
-                    return $content
+                    return $dscContent
                 }
                 return $returnValue
             } -ArgumentList @($batch, $PSScriptRoot, $GlobalAdminAccount, $ApplicationId, $TenantId, $CertificateThumbprint, $organization) | Out-Null
@@ -518,12 +521,12 @@ function Export-TargetResource
                 {
                     $partialResult = Receive-Job -Name $job.name
                     $result += $partialResult
-                    Remove-Job -Name $job.name
+                    Remove-Job -Name $job.name | Out-Null
                     $jobsCompleted++
                 }
                 elseif ($job.JobStateInfo.State -eq 'Failed')
                 {
-                    Remove-Job -Name $job.name
+                    Remove-Job -Name $job.name | Out-Null
                     Write-Warning "{$($job.name)} failed"
                     break
                 }
@@ -548,6 +551,7 @@ function Export-TargetResource
     }
     catch
     {
+        Write-Host $Global:M365DSCEmojiRedX
         try
         {
             Write-Verbose -Message $_